+++ packages/template-retail-react-app/app/utils/product-utils.js
@@ -170,7 +170,16 @@ export const getDecoratedVariationAttributes = (product, opts = {}) => {
  */
 export const findLowestPrice = (product) => {
     if (!product) return
-    const array = product.variants ?? [product]
+
+    // Look at all of the variants, only if it's a master product.
+    // i.e. when a shopper has narrowed down to a variant, do not look into other variants
+    const isMaster = product.productType?.master || product.type?.master
+    if (isMaster && !product.variants) {
+        console.warn(
+            'Expecting `product.variants` to exist. For more accuracy, please tweak your API request to ask for the variants details.'
+        )
+    }
+    const array = isMaster && product.variants ? product.variants : [product]
 
     return array.reduce(
         (prev, data) => {